merkle: with_depth_and_leaf to create a tree
I'm still not convinced with the associate function
name itself, but this will initialize and calculate
merkle root with the provided depth of the table and
the initial leaf hash, which will be set for all
the leaves.

By the way, this is a good way to generate the actual
tree with the resonable computation because each layers
contains the same hash and makes the O(depth) runtime
complexity.