merkle: MerkleProof type
Instead of returning MerkleProofIter,
returning new MerkleProof type value
by the proof() function, which only
contains the data needed to do the
merkle proof, e.g. node kind and the
sibling hash.  This makes sense from
the actual blockchain operations,
which constructs the merkle proof from
the data and calculate the root hash
out of it.

It also doesn't have any link to the
MerkleTree type itself, as that's
resides in the blockchain node, not
the client/wallet.