merkle: with_leaves constructor
This is the default way to create
the MerkleTree.  The goal is to
make it FromIterator trait but that
requires the reverse ordering of
the Vec, leaves to the root instead
of the other way around, just like
the official merkle crate.

Until then, I'll stick with this
constructor.

The previous one was dropped, as
that's mainly for the debugging
purpose.